+++ cmd/backup.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"archive/tar"
 	"compress/gzip"
-	"io"
+	"fmt"
 	"os"
 	"path/filepath"
 	"time"
@@ -14,15 +14,20 @@ import (
 	"github.com/EbonJaeger/mcsmanager/tmux"
 )
 
+type BackupFlags struct {
+	Level int `short:"l" long:"level" desc:"Set the compression format to use; 0: no compression; 1: gzip"`
+}
+
 // Backup archives the Minecraft server files.
 var Backup = cmd.Sub{
 	Name:  "backup",
 	Alias: "b",
-	Short: "Backup all server files into a .tar.gz archive",
+	Short: "Backup all server files into a tar archive with optional compression",
+	Flags: &BackupFlags{},
 	Run:   ArchiveServer,
 }
 
-// ArchiveServer adds all directories and files of the server into a Gzip'd tar archive.
+// ArchiveServer adds all directories and files of the server into a tar archive with optional compression.
 func ArchiveServer(root *cmd.Root, c *cmd.Sub) {
 	prefix, err := root.Flags.(*GlobalFlags).GetPathPrefix()
 	if err != nil {
@@ -50,121 +55,77 @@ func ArchiveServer(root *cmd.Root, c *cmd.Sub) {
 	// Check if the backup directory exists
 	if _, err := os.Stat(backupDir); os.IsNotExist(err) { // Dir does not exist
 		Log.Infoln("Backup directory does not exist! Creating it...")
-		err = os.Mkdir(backupDir, 0755)
-		if err != nil {
-			Log.Fatalln("Unable to create backups directory:", err.Error())
+		if err = os.Mkdir(backupDir, 0755); err != nil {
+			Log.Fatalln("Unable to create backups directory: %s\n", err)
 		}
 		Log.Goodln("Backup directory created!")
 	}
 
 	// Check for backups that are too old
-	pruned, err := mcsmanager.RemoveOldFiles(backupDir, conf.BackupSettings.MaxAge)
-	if err != nil {
-		Log.Fatalf("Unable to remove old backups: %s\n", err.Error())
-	}
-	if pruned > 0 {
-		Log.Infof("Removed %d archive(s) due to age.\n", pruned)
+	if pruned, err := mcsmanager.PruneOld(backupDir, conf.BackupSettings.MaxAge); err == nil {
+		if pruned > 0 {
+			Log.Infof("Removed %d archive(s) due to age.\n", pruned)
+		}
+	} else {
+		Log.Fatalf("Unable to remove old backups: %s\n", err)
 	}
 
 	// Check for too many backups
-	pruned, err = mcsmanager.RemoveTooManyFiles(backupDir, conf.BackupSettings.MaxBackups)
-	if err != nil {
-		Log.Fatalf("Unable to remove old backups: %s\n", err.Error())
-	}
-	if pruned > 0 {
-		Log.Infof("Removed %d archive(s) because over backup limit.\n", pruned)
+	if pruned, err := mcsmanager.Prune(backupDir, conf.BackupSettings.MaxBackups); err == nil {
+		if pruned > 0 {
+			Log.Infof("Removed %d archive(s) because over backup limit.\n", pruned)
+		}
+	} else {
+		Log.Fatalf("Unable to remove old backups: %s\n", err)
 	}
 
 	// Create archive file
-	tarFile, err := createArchiveFile(backupDir)
-	defer tarFile.Close()
+	level := c.Flags.(*BackupFlags).Level
+	tarFile, err := createArchive(backupDir, level)
 	if err != nil {
-		Log.Fatalf("Error while adding files to archive: %s\n", err.Error())
+		Log.Fatalf("Error while adding files to archive: %s\n", err)
 	}
+	defer tarFile.Close()
 
 	// Create our file writers
-	fileWriter := gzip.NewWriter(tarFile)
-	defer fileWriter.Close()
-	tarFileWriter := tar.NewWriter(fileWriter)
-	defer tarFileWriter.Close()
-
-	src, err := os.Open(".")
-	defer src.Close()
-	if err != nil {
-		Log.Fatalf("Error while adding files to archive: %s\n", err.Error())
+	var w *tar.Writer
+	switch level {
+	case 0:
+		w = tar.NewWriter(tarFile)
+	case 1:
+		compressor, err := gzip.NewWriterLevel(tarFile, gzip.BestCompression)
+		if err != nil {
+			Log.Fatalln("Failed to create archive compressor: %s\n", err)
+		}
+		defer compressor.Close()
+		w = tar.NewWriter(compressor)
 	}
+	defer w.Close()
 
-	// Archive the server directory recursively
-	archiveDir(src, tarFileWriter)
-	Log.Goodln("Server file archive created!")
+	// Add all of the server files to the archive
+	start := time.Now()
+	err = mcsmanager.Archive(prefix, w, "backups")
+	diff := time.Since(start)
+
+	if err != nil {
+		Log.Errorf("Error adding files to archive: %s\n", err)
+	}
+	Log.Goodf("Server backup archive created in %s\n", diff.String())
 }
 
-func createArchiveFile(dir string) (*os.File, error) {
+func createArchive(dir string, level int) (*os.File, error) {
 	currentTime := time.Now()
 	timeStr := currentTime.Format("2006-01-02T15:04:05-0700") // ISO-8601 format
-	tarPath := filepath.Join(dir, timeStr+".tar.gz")
 
-	return os.Create(tarPath)
-}
-
-func archiveDir(dir *os.File, w *tar.Writer) {
-	files, err := dir.Readdir(-1)
-	if err != nil {
-		Log.Errorf("Error read files in directory: %s\n", err.Error())
+	var path string
+	switch level {
+	case 0: // No compression
+		path = filepath.Join(dir, timeStr+".tar")
+	case 1: // gzip compression
+		path = filepath.Join(dir, timeStr+".tar.gz")
+	default:
+		return nil, fmt.Errorf("compression level not supported: %d", level)
 	}
 
-	// Iterate through all files
-	for _, fileInfo := range files {
-		// Exclude backups directory
-		if fileInfo.Name() == "backups" {
-			continue
-		}
-
-		if fileInfo.IsDir() { // File is actually a directory
-			nestedDir, err := os.Open(dir.Name() + string(filepath.Separator) + fileInfo.Name())
-			defer nestedDir.Close()
-			if err != nil {
-				Log.Errorf("Error opening directory for archiving: %s\n", err.Error())
-				continue
-			}
-
-			// Write directory header to archive
-			header, _ := tar.FileInfoHeader(fileInfo, "")
-			header.Name = dir.Name() + string(filepath.Separator) + fileInfo.Name()
-			err = w.WriteHeader(header)
-			if err != nil {
-				Log.Errorf("Error adding directory to archive: %s\n", err.Error())
-			}
-
-			// Recurse and archive everything else in this directory
-			archiveDir(nestedDir, w)
-		} else { // File is a file, archive it normally
-			archiveFile(dir, fileInfo, w)
-		}
-	}
-}
-
-func archiveFile(dir *os.File, fi os.FileInfo, w *tar.Writer) {
-	file, err := os.Open(dir.Name() + string(filepath.Separator) + fi.Name())
-	defer file.Close()
-	if err != nil {
-		Log.Errorf("Error writing file to archive: %s\n", err.Error())
-	}
-
-	// Create tar header
-	header := new(tar.Header)
-	header.Name = file.Name()
-	header.Size = fi.Size()
-	header.Mode = int64(fi.Mode())
-	header.ModTime = fi.ModTime()
-
-	err = w.WriteHeader(header)
-	if err != nil {
-		Log.Errorf("Error writing file to archive: %s\n", err.Error())
-	}
-
-	_, err = io.Copy(w, io.Reader(file))
-	if err != nil {
-		Log.Errorf("Error writing file to archive: %s\n", err.Error())
-	}
+	return os.Create(path)
 }